+++ packages/build-tools/src/config/babel.base.config.js
@@ -10,5 +10,7 @@ module.exports = {
     '@babel/plugin-syntax-jsx',
     '@babel/proposal-class-properties',
     '@babel/proposal-object-rest-spread',
+    '@babel/plugin-proposal-nullish-coalescing-operator',
+    '@babel/plugin-proposal-optional-chaining',
   ],
 };