+++ packages/RangeFilter/CHANGELOG.md
@@ -3,6 +3,65 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.8.0-alpha.60](https://github.com/thebyte9/blaze-components-react/compare/v0.8.0-alpha.59...v0.8.0-alpha.60) (2021-11-30)
+
+**Note:** Version bump only for package @blaze-react/range-filter
+
+
+
+
+
+# [0.8.0-alpha.59](https://github.com/thebyte9/blaze-components-react/compare/v0.8.0-alpha.58...v0.8.0-alpha.59) (2021-11-29)
+
+**Note:** Version bump only for package @blaze-react/range-filter
+
+
+
+
+
+# [0.8.0-alpha.58](https://github.com/thebyte9/blaze-components-react/compare/v0.8.0-alpha.57...v0.8.0-alpha.58) (2021-11-29)
+
+**Note:** Version bump only for package @blaze-react/range-filter
+
+
+
+
+
+# [0.8.0-alpha.57](https://github.com/thebyte9/blaze-components-react/compare/v0.8.0-alpha.56...v0.8.0-alpha.57) (2021-11-29)
+
+**Note:** Version bump only for package @blaze-react/range-filter
+
+
+
+
+
+# [0.8.0-alpha.54](https://github.com/thebyte9/blaze-components-react/compare/v0.8.0-alpha.53...v0.8.0-alpha.54) (2021-11-29)
+
+**Note:** Version bump only for package @blaze-react/range-filter
+
+
+
+
+
+# [0.8.0-alpha.53](https://github.com/thebyte9/blaze-components-react/compare/v0.8.0-alpha.52...v0.8.0-alpha.53) (2021-11-29)
+
+**Note:** Version bump only for package @blaze-react/range-filter
+
+
+
+
+
+# [0.8.0-alpha.52](https://github.com/thebyte9/blaze-components-react/compare/v0.8.0-alpha.49...v0.8.0-alpha.52) (2021-11-29)
+
+
+### Bug Fixes
+
+* range number formatting ([#582](https://github.com/thebyte9/blaze-components-react/issues/582)) ([8839ee6](https://github.com/thebyte9/blaze-components-react/commit/8839ee6bd4e9cc502d57d71b8acf3cda636d0266))
+
+
+
+
+
 # [0.8.0-alpha.46](https://github.com/thebyte9/blaze-components-react/compare/v0.8.0-alpha.45...v0.8.0-alpha.46) (2021-10-06)
 
 **Note:** Version bump only for package @blaze-react/range-filter